static/images/From scratch/requirements.txt
flask
pytesseract
gtts
OpenCV
tensorflow
pillow
pygame
torch
numpy
pandas
